Some refactor to the code
