+++ src/components/AppLayout/Header/components/ProviderDetails/UserDetails.tsx
@@ -1,10 +1,10 @@
+import * as React from 'react'
+import styled from 'styled-components'
+import classNames from 'classnames'
 import { makeStyles } from '@material-ui/core/styles'
 import Dot from '@material-ui/icons/FiberManualRecord'
-import classNames from 'classnames'
-import * as React from 'react'
 import { EthHashInfo, Identicon, Card } from '@gnosis.pm/safe-react-components'
 import { createStyles } from '@material-ui/core'
-import styled from 'styled-components'
 
 import Spacer from 'src/components/Spacer'
 import Block from 'src/components/layout/Block'
